requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.40.0
sqlalchemy==2.0.37
pydantic==2.12.5
python-multipart==0.0.20
pillow==10.4.0
openai==1.61.0
python-dotenv==1.2.1
requests==2.32.3
fastmcp==2.14.4
mcp==1.26.0
pypdf==6.0.0
python-docx==1.2.0
